+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @attn-protocol/framework
 
+## 0.4.0
+
+### Minor Changes
+
+- test
+
 ## 0.3.0
 
 ### Minor Changes